+++ src/commands/playground.ts
@@ -6,12 +6,7 @@ import { EXTENSION_ID, FS_SCHEME, PLAYGROUND_JSON_FILE } from "../constants";
 import { IPlaygroundJSON } from "../interfaces/IPlaygroundJSON";
 import { Gist } from "../store";
 import { newGist } from "../store/actions";
-import {
-  byteArrayToString,
-  closeGistFiles,
-  fileNameToUri,
-  stringToByteArray
-} from "../utils";
+import { byteArrayToString, closeGistFiles, fileNameToUri, stringToByteArray } from "../utils";
 import { PlaygroundWebview } from "../webView";
 import { addPlaygroundLibraryCommand } from "./addPlaygroundLibraryCommand";
 import { getCDNJSLibraries } from "./cdnjs";