+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## v0.0.34
+## v0.0.34 (12/28/2019)
 
+- Introduced the `Duplicate Gist` command to the `Gists` tree view, which allows you to create new gists, based on the contents of an existing one.
 - Added multi-select support to the `Gists` tree view, and added support for the following commands: deleting gists, unstarring gists, deleting gist files, and unfollowing users.
 - The `Starred Gists` node is now hidden in the tree view unless you actually have any starred gists. Additionally, when displayed, this node is collapsed by default.